requirements.txt
psycopg2-binary==2.9.1
web3==5.31.3